requirements.txt
asgiref==3.6.0
Django==4.1.4
django-dynamic-preferences==1.14.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-taggit==3.1.0
persisting-theory==1.0
Pillow==9.3.0
psycopg2==2.9.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
